Create the cartesian product of two sets

Given two sets creates new sets with one element of each set

cartesian(object, set1, set2, name = NULL, ...)

# S3 method for TidySet
cartesian(
  object,
  set1,
  set2,
  name = NULL,
  keep = TRUE,
  keep_relations = keep,
  keep_elements = keep,
  keep_sets = keep,
  ...
)

Arguments

object
A TidySet object.

set1, set2
The name of the sets to be used for the cartesian product

name
The name of the new set.

...
Other arguments passed down if possible

keep
A logical value if you want to keep

keep_relations
A logical value if you wan to keep old relations

keep_elements
A logical value if you wan to keep old elements

keep_sets
A logical value if you wan to keep old sets

Value

A TidySet object with the new set

See also

Examples

relations <- data.frame(sets = c(rep("a", 5), "b"),
                        elements = letters[seq_len(6)])
TS <- tidySet(relations)
cartesian(TS, "a", "b")

#>    elements  sets fuzzy
#> 1         a     a     1
#> 2         a a⨯b_1     1
#> 3         a   a⨯b     1
#> 4         b     a     1
#> 5         b   a⨯b     1
#> 6         b a⨯b_2     1
#> 7         c   a⨯b     1
#> 8         c     a     1
#> 9         c a⨯b_3     1
#> 10        d     a     1
#> 11        d a⨯b_4     1
#> 12        d   a⨯b     1
#> 13        e   a⨯b     1
#> 14        e a⨯b_5     1
#> 15        e     a     1
#> 16        f     b     1
#> 17        f a⨯b_1     1
#> 18        f a⨯b_2     1
#> 19        f a⨯b_3     1
#> 20        f a⨯b_4     1
#> 21        f a⨯b_5     1
#> 22        f   a⨯b     1