Pull from a TidySet

Use pull to extract the columns of a TidySet object. You can use activate with filter or use the specific function. The S3 method filters using all the information on the TidySet.

# S3 method for TidySet
pull(.data, var = -1)

pull_set(.data, var)

pull_element(.data, var)

pull_relation(.data, var)

Arguments

.data
The TidySet object

var
The literal variable name, a positive integer or a negative integer column position.

Value

A TidySet object

See also

Examples

relations <- data.frame(sets = c(rep("a", 5), "b", rep("a2", 5), "b2"),
                        elements = rep(letters[seq_len(6)], 2),
                        fuzzy = runif(12))
a <- tidySet(relations)
a <- mutate_element(a, type = c(rep("Gene", 4), rep("lncRNA", 2)))
pull(a, type)

#>  [1] "Gene"   "Gene"   "Gene"   "Gene"   "Gene"   "Gene"   "Gene"   "Gene"  
#>  [9] "lncRNA" "lncRNA" "lncRNA" "lncRNA"

# Equivalent to pull_relation
b <- activate(a, "relations")
pull_relation(b, elements)

#>  [1] a b c d e f a b c d e f
#> Levels: a b c d e f

pull_element(b, elements)

#> [1] a b c d e f
#> Levels: a b c d e f

# Filter element
pull_element(a, type)

#> [1] "Gene"   "Gene"   "Gene"   "Gene"   "lncRNA" "lncRNA"

# Filter sets
pull_set(a, sets)

#> [1] a  b  a2 b2
#> Levels: a a2 b b2